Compute improvements along with simplification from chains.
The solver computes simplifications in cases where there is only one
possible way to solve a given constraint.  However, it was not
previously also computing an improvement in those cases.  This is
captured in `impr9`:

    F t u | t ~> u.
    F t (S t) if O t; F t u fails.

    F t u, C u?

The solver should simplify the `F t u` constraint to `O t`, but in doing
so must also introduce the improvement `u :-> S t`.